chore: Remove root user from authentik.yml service
This commit removes the `user: root` configuration from the `authentik.yml` service in the Docker Compose file. This change is made to prevent the worker from fixing the permissions on the mounted folders. It is recommended to ensure that the folders have the correct UID/GID (1000:1000 by default) when removing this configuration.

Recent user commits:
- Update docker-compose-arr.yml to include container names for services
- Update .env.sample with Redis and Postgres versions
- Add metrics endpoint for Prometheus monitoring
- Update docker-compose-arr.yml to include profiles for core and all services
- Update traefik.yml to configure Traefik for HTTPS and HTTP redirection
- Update healthcheck test in authentik.yml to use AUTHENTIK_PG_DB and AUTHENTIK_PG_USER
- Update authentik.yml with email configuration
- Add arr/authentik.yml to docker-compose-arr.yml