chore: Update qBittorrent port in gluetun.yml and qbittorrent.yml
This commit updates the qBittorrent port in the `gluetun.yml` and `qbittorrent.yml` Docker Compose files. The previous port was set to `${QBIT_WEBUI_PORT:-8085}`, but it has been changed to `${QBIT_WEBUI_PORT:-8080}` to align with the desired configuration for exposing qBittorrent through the Docker Host LAN IP. This change ensures that the correct port is used for accessing qBittorrent.

Recent user commits:
- Update traefik.yml to configure Traefik for HTTPS and HTTP redirection
- Remove root user from authentik.yml service
- Update docker-compose-arr.yml to include container names for services
- Update .env.sample with Redis and Postgres versions
- Add metrics endpoint for Prometheus monitoring
- Update docker-compose-arr.yml to include profiles for core and all services
- Update healthcheck test in authentik.yml to use AUTHENTIK_PG_DB and AUTHENTIK_PG_USER
- Update authentik.yml with email configuration
- Add arr/authentik.yml to docker-compose-arr.yml