chore: Update Traefik configuration for qBittorrent in gluetun.yml
This commit updates the Traefik configuration in the `gluetun.yml` Docker Compose file to expose qBittorrent through the Docker Host LAN IP. The `traefik.http.routers.qbit-gluetun` and `traefik.http.routers.qbit-gluetun-secure` labels are added to enable routing and secure access to qBittorrent. The port is set to `8080` and HTTPS is enforced for secure connections.