chore: Update profiles in Docker Compose files for ARR services
This commit updates the profiles in the Docker Compose files for ARR services. Specifically, the profiles in the `prowlarr-exporter.yml`, `prowlarr.yml`, `radarr-exporter.yml`, `sonarr-exporter.yml`, and `stash.yml` files have been modified. The changes involve removing the `media` profile and adding the `arrs` profile in some cases, while in others, the `apps` profile is added. These updates ensure that the correct profiles are used for the ARR services.