chore: Update Traefik configuration for authentik in authentik.yml
This commit updates the Traefik configuration in the `authentik.yml` Docker Compose file. It adds the `traefik.http.middlewares.sslheader.headers.customrequestheaders.X-Forwarded-Proto=https` label to enable HTTPS redirection. The `traefik.http.routers.authentik-secure.tls` label is commented out to disable TLS. The port for the `authentik` service is set to `9000`. These changes ensure that the Traefik configuration is correctly updated for the authentik service.