chore: Remove commented out API_KEY in flaresolverr.yml
This commit removes the commented out API_KEY line in the flaresolverr.yml Docker Compose file. The line was previously disabled by adding a '#' at the beginning of the line. This change ensures that the API_KEY environment variable is not set to a default value and allows for custom configuration of the API_KEY.