chore: Update Grafana provisioning directory in grafana.yml
This commit updates the Grafana provisioning directory in the grafana.yml file. The source path for the provisioning files is changed from "$DOCKERDIR/appdata/grafana" to "$DOCKERDIR/appdata/grafana/provisioning". This change ensures that the provisioning files are correctly located and can be used for proper configuration and functionality of Grafana.