chore: Update Grafana provisioning directory and add volume for grafana_data
This commit updates the Grafana provisioning directory in the grafana.yml file and adds a volume for `grafana_data`. The provisioning directory is changed to "$DOCKERDIR/appdata/grafana/provisioning" to ensure proper configuration and functionality of Grafana. The volume `grafana_data` is added to persist data for Grafana. These changes improve the management and functionality of Grafana.