chore: Update Grafana and Prometheus provisioning directories
This commit updates the Grafana and Prometheus provisioning directories. The Grafana provisioning directory is changed to "$DOCKERDIR/appdata/grafana/provisioning" and the Prometheus provisioning directory is changed to "$DOCKERDIR/appdata/prometheus/provisioning". These changes ensure that the provisioning files are correctly located and can be used for proper configuration and functionality of Grafana and Prometheus.