chore: Update Traefik configuration for TLS settings
This commit updates the Traefik configuration in the traefik.yml file. The TLS settings for the websecure entrypoint are enabled by uncommenting the corresponding lines. Additionally, the TLS certificate resolver is set to "dns-cloudflare" and the main and SANs domains are configured with the value of $DOMAINNAME_HS. These changes improve the security and functionality of Traefik.